% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{ecf}
\alias{ecf}
\title{Empirical characteristic function}
\usage{
ecf(t, samp)
}
\arguments{
\item{t}{d-variate vector where the ecf will be evaluated}

\item{samp}{nxd matrix with the n samples}
}
\value{
A complex value
}
\description{
It computes the d-dimensional empirical characteristic function
}
\examples{

library(mvdeconvolution)
library(reshape2)
library(ggplot2)

## Example 1
## Real and imag. parts of exponential dist.
samp <- rexp(60)
t <- seq(-3, 3, len = 150)
vals <- sapply(t, ecf, samp)
real <- Re(vals)
imag <- Im(vals)
## ## Plot values. Uncomment below
## df <- data.frame(t, real, imag)
## df.molten <- melt(df, id = "t")
## head(df.molten)
## qplot(t, value, data = df.molten, geom = 'line',
##       facets = . ~ variable)

## Example 2
## Real part of bivariate normal emp. ch. fnc.
samp <- cbind(rnorm(100), rnorm(100))
t1 <- seq(-3, 3, len = 150)
t2 <- t1
t <- expand.grid(t1, t2)
ecf.vals <- apply(t, 1, ecf, samp)
re.vals <- matrix(Re(ecf.vals), 150, 150)
persp(t1, t2, re.vals)
}

