% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{integrand}
\alias{integrand}
\title{Integrand for Fourier transform}
\usage{
integrand(t, EMethod, error, sigE, truncate, h, samp, columns, Kkernel)
}
\arguments{
\item{t}{A bivariate vector.}

\item{EMethod}{Method used to approximate the characteristic function of
the error. 1 or 2 if ecf of the char. fnc. of the kde is used, respectively.}

\item{error}{Matrix with 2 columns, either with the pure sample of the error, or with the
differences obtained from the panel data structure.}

\item{sigE}{Bandwidth (2x2) matrix for the error. It is NULL by default
and it is required only when EMethod = 2.}

\item{truncate}{NULL by default. If the approximated characteristic function
of the error is smaller than truncate, it will make the integrand zero.
This provides numerical stability.}

\item{h}{Bivariate vector with the bandwidth parameters.}

\item{samp}{2-column matrix with the contaminated sample.}

\item{columns}{Number of columns of the panel data. If there is one column,
it assumes that instead panel data, a pure sample of the error is provided.}

\item{Kkernel}{Kernel to be used. See 'ker' function.}
}
\value{
A complex value.
}
\description{
This is the integrand in the bivariate deconvolution formula, except
for the exp(-itx) term.
}
\seealso{
kerdecon
}

