% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{ker}
\alias{ker}
\title{FT of product kernels}
\usage{
ker(t, kernel)
}
\arguments{
\item{t}{d-dimensional vector to evaluate the FT.}

\item{kernel}{Value from 1 to 5. 1, sinc kernel;
2, VP kernel;
3, kernel whose FT is prop. to triweight kernel;
4, kernel whose FT is prop. to Tricube kernel, and
5, flat-top kernel.}
}
\description{
It computes the Fourier transform of several kernels
with support in -1 to 1.
}
\examples{

library(mvdeconvolution)
library(reshape2)
library(ggplot2)

## Example for the univariate case
## Plot all the possible kernels
t <- seq(-1.2, 1.2, length.out = 150)
sinc <- sapply(t, ker, 1)
VP <- sapply(t, ker, 2)
triw <- sapply(t, ker, 3)
tric <- sapply(t, ker, 4)
flat <- sapply(t, ker, 5)

## ## Plot. Uncomment code below
## df <- data.frame(t, sinc, VP, triw, tric, flat)
## df.molten <- melt(df, id = "t", variable.name = "kernel")
## qplot(t, value, data = df.molten,
##       geom = "line", colour = kernel)

}

